make it so on new positive search but negative search, defaults to everything
example: `music \!the`, before this would match nothing since there is
no term to match, only a term to not match.

Before you would have to do something like `music \!the .` to fix it;
Now the program assumes if there is only negation filters, by default
to include it.

Arguable which is the more logical solution, but i'll just use this for
now