fix --live quote backspace deletion & wrong reading of bytes
when backspacing in --live mode, if it deleted a quote the variable
that held whether a quote was finished or not was not updated.

also the stdin read 1 byte each time and arrow keys were broken up into
"\x1B" "[" "A|B|C|D" which resulted in them not being handled propely.
Because "\x1B" is the code for the escape, and my program recognizes
that as a sign to exit, the arrow keys caused the program to exit.

fixed by reading 3 bytes instead and truncating null bytes